+++ packages/core/expressions/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.49](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.4.48...@kong-ui-public/expressions@0.4.49) (2024-10-15)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.4.48](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.4.47...@kong-ui-public/expressions@0.4.48) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
+## [0.4.47](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.4.46...@kong-ui-public/expressions@0.4.47) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/expressions
+
+
+
+
+
 ## [0.4.46](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/expressions@0.4.45...@kong-ui-public/expressions@0.4.46) (2024-10-09)
 
 **Note:** Version bump only for package @kong-ui-public/expressions