+++ packages/entities/entities-certificates/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.32](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.31...@kong-ui-public/entities-certificates@3.6.32) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.6.31](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.30...@kong-ui-public/entities-certificates@3.6.31) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
+## [3.6.30](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.29...@kong-ui-public/entities-certificates@3.6.30) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-certificates
+
+
+
+
+
 ## [3.6.29](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-certificates@3.6.28...@kong-ui-public/entities-certificates@3.6.29) (2024-10-09)
 
 **Note:** Version bump only for package @kong-ui-public/entities-certificates