+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.16...@kong-ui-public/entities-gateway-services@3.8.17) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.8.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.15...@kong-ui-public/entities-gateway-services@3.8.16) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
+## [3.8.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.14...@kong-ui-public/entities-gateway-services@3.8.15) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
 ## [3.8.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.8.13...@kong-ui-public/entities-gateway-services@3.8.14) (2024-10-09)
 
 **Note:** Version bump only for package @kong-ui-public/entities-gateway-services