+++ packages/entities/entities-vaults/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.13.24](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.23...@kong-ui-public/entities-vaults@3.13.24) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-vaults
+
+
+
+
+
+## [3.13.23](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.22...@kong-ui-public/entities-vaults@3.13.23) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-vaults
+
+
+
+
+
+## [3.13.22](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.21...@kong-ui-public/entities-vaults@3.13.22) (2024-10-14)
+
+**Note:** Version bump only for package @kong-ui-public/entities-vaults
+
+
+
+
+
 ## [3.13.21](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.13.20...@kong-ui-public/entities-vaults@3.13.21) (2024-10-09)
 
 **Note:** Version bump only for package @kong-ui-public/entities-vaults